+++ src/models/index.ts
@@ -1,4 +1,7 @@
 export * from './CategoryModel';
+export * from './CreditModel';
 export * from './FeaturedMovieModel';
+export * from './MovieDetailsModel';
+export * from './MovieListItemModel';
 export * from './MovieListModel';
 export * from './MovieModel';
